backend/requirements.txt
# Backend Dependencies
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
supabase==2.6.0
pdfplumber==0.11.8
python-dotenv==1.2.1
langchain==0.2.17
langchain-text-splitters==0.2.4
langchain-community==0.2.19
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
inngest>=0.3.0